// src/core/domain/failure-codes.enum.ts
export enum FailureCode {
  INSUFFICIENT_FUNDS = 'INSUFFICIENT_FUNDS',
  CURRENCY_MISMATCH = 'CURRENCY_MISMATCH',
  INVALID_REFERENCE = 'INVALID_REFERENCE',
  NEGATIVE_BALANCE_ATTEMPT = 'NEGATIVE_BALANCE_ATTEMPT',
  BUSINESS_RULE_VIOLATION = 'BUSINESS_RULE_VIOLATION',
  UNKNOWN_ERROR = 'UNKNOWN_ERROR',
}